Add new dashboard layout and update various pages
This commit introduces a new layout for the dashboard that includes a side nav. It also brings updates to the page components such as invoices, customers, overview and its loading page. In addition to this, navigation links are enhanced with a conditional class. Newly introduced fonts are applied in app/page and layout files. Moreover, the app/page file now incorporates hero images. A new css file is also added for homepage styling.